+++ packages/codemirror-block-editor/CHANGELOG.md
@@ -4,4 +4,5 @@
 
 - Basic indentation with Tab and Shift-Tab
 - Increase and decrease levels of lines
+- Increase block level with constraints
 - Initial version